requirements.txt
selenium
python-gitlab
python-dotenv
pyinstaller
